chore: upgrade to ESLint 10 with flat config
