+++ packages/package.json
@@ -7,9 +7,8 @@
     "npm": ">=11.0.0"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^8.6.0",
-    "@typescript-eslint/parser": "^8.6.0",
-    "@wordpress/eslint-plugin": "^21.1.0",
+    "@wordpress/eslint-plugin": "^25.8.0",
+    "typescript-eslint": "^8.65.0",
     "tsup": "^8.3.5"
   },
   "scripts": {
