+++ packages/packages/tools/eslint-plugin-editor/package.json
@@ -45,7 +45,7 @@
   },
   "peerDependencies": {
     "@typescript-eslint/parser": "^8.6.0",
-    "eslint": "^8.57.1",
+    "eslint": "^8.57.1 || ^9.0.0 || ^10.0.0",
     "typescript": "^5.7.2"
   }
 }
